updated with working edit review function - title only for now
