+++ controllers/dashboard-routes.js
@@ -45,50 +45,4 @@ router.get('/', withAuth, (req, res) => {
       });
   });
 
-  router.get('/edit/:id', withAuth, (req, res) => {
-    Post.findByPk(req.params.id, {
-    attributes: [
-        'id',
-        'review_title',
-        'review_text',
-        'is_public',
-        'comments_enabled',
-        'created_at'
-    ],
-    include: [
-      {
-        model: Book,
-        attributes:['title']
-      },
-        {
-        model: Comment,
-        attributes: ['id', 'comment_text', 'review_id', 'user_id', 'created_at'],
-        include: {
-            model: User,
-            attributes: ['username']
-        }
-        },
-        {
-        model: User,
-        attributes: ['username']
-        }
-    ]
-    })
-    .then(dbReviewData => {
-        if (dbReviewData) {
-        const review = dbReviewData.get({ plain: true });
-        
-        res.render('edit-review', {
-            review,
-            loggedIn: true
-        });
-        } else {
-        res.status(404).end();
-        }
-    })
-    .catch(err => {
-        res.status(500).json(err);
-    });
-    });
-
   module.exports = router;
