+++ controllers/index.js
@@ -4,6 +4,7 @@ const apiRoutes = require('./api/');
 const homeRoutes = require('./home-routes.js');
 const dashboardRoutes = require('./dashboard-routes.js');
 const createReviewRoutes = require('./create-review-routes');
+const editReviewRoutes = require('./edit-review-routes');
 
 
 
@@ -12,5 +13,6 @@ router.use('/api', apiRoutes);
 router.use('/', homeRoutes);
 router.use('/dashboard', dashboardRoutes);
 router.use('/create-review', createReviewRoutes);
+router.use('/edit-review',  editReviewRoutes);
 
 module.exports = router;
