+++ public/javascript/edit-review.js
@@ -1,11 +1,11 @@
 async function editFormHandler(event) {
     event.preventDefault();
   
-    const title = document.querySelector('input[name="review_title"]').value.trim();
+    const title = document.querySelector('input[name="review-title"]').value.trim();
     const id = window.location.toString().split('/')[
     window.location.toString().split('/').length - 1
     ];
-    const response = await fetch(`/api/posts/${id}`, {
+    const response = await fetch(`/api/reviews/${id}`, {
       method: 'PUT',
       body: JSON.stringify({
         title
@@ -20,4 +20,4 @@ async function editFormHandler(event) {
 
 }
 
-document.querySelector('.edit-post-form').addEventListener('submit', editFormHandler);
+document.querySelector('.edit-review-form').addEventListener('submit', editFormHandler);
