+++ views/partials/review-info-dash.handlebars
@@ -3,6 +3,6 @@
         {{review_title}}
     </h3>
     <div class='review-details'>
-        Created: {{format_date created_at}} | <span class='delete-btn'>Delete</span>
+        Created: {{format_date created_at}} | <a href='/edit-review/{{id}}'>Edit</a> | <span class='delete-btn' id='r{{id}}'>Delete</span>
     </div>
 </article>
